+++ src/Entity/Statement.php
@@ -33,7 +33,7 @@ class Statement
     private ?int $amount = 0;
 
     #[ORM\Column]
-    private ?int $cost = 0;
+    private ?float $cost = 0;
 
     public function getId(): ?int
     {
@@ -100,12 +100,12 @@ class Statement
         return $this;
     }
 
-    public function getCost(): ?int
+    public function getCost(): ?float
     {
         return $this->cost;
     }
 
-    public function setCost(int $cost): self
+    public function setCost(float $cost): self
     {
         $this->cost = $cost;
